6 complete basic tests

+++ bash/refresh.sh
@@ -1,20 +1,24 @@
 #!/usr/bin/env bash
 
+# TODO - document this file
 #rm -f ./database/database.sqlite
 #touch ./database/database.sqlite
 
 CLEAR_CACHE=no
 DELETE_DB=no
+TEST_MIGRATIONS=no
 
-while getopts ":c:d:" opt
+while getopts ":c:d:t:" opt
    do
      case $opt in
-        c ) CLEAR_CACHE=$OPTARG;; # update in in vars
-        d ) DELETE_DB=$OPTARG;; # update in in vars
+        c ) CLEAR_CACHE=$OPTARG;;
+        d ) DELETE_DB=$OPTARG;;
+        t ) TEST_MIGRATIONS=$OPTARG;;
      esac
 done
 
-if [ ${CLEAR_CACHE} = 'no' ]
+# clearing cache?
+if [ ${CLEAR_CACHE} = 'no' ] && [ ${TEST_MIGRATIONS} = 'no' ]
 then
     echo "Cache was not cleared"
 else
@@ -22,7 +26,8 @@ else
     ./bash/clear.sh
 fi
 
-if [ ${DELETE_DB} = 'no' ]
+# deleting old database?
+if [ ${DELETE_DB} = 'no' ] && [ ${TEST_MIGRATIONS} = 'no' ]
 then
     echo "Database was not deleted"
     php artisan migrate:reset
@@ -32,4 +37,14 @@ else
     touch ./database/database.sqlite
 fi
 
+# running migrations
 php artisan migrate
+
+# running test migrations?
+if [ ${TEST_MIGRATIONS} = 'no' ]
+then
+    echo "Test migrations were not ran"
+else
+    echo "Running testing migrations"
+    php artisan migrate --path /database/migrations/test
+fi
